@@ -6,7 +6,7 @@
   "scripts": {
     "psy-quiz-backend": "NX_NO_CLOUD=true && NX_SKIP_NX_CACHE=true && nx serve psy-quiz-backend --configuration development",
     "psy-quiz-backend-build": "NX_NO_CLOUD=true && NX_SKIP_NX_CACHE=true && nx build psy-quiz-backend --configuration development",
-    "psy-quiz-frontend": "nx serve psy-quiz-frontend",
+    "psy-quiz-frontend": "NX_NO_CLOUD=true && NX_SKIP_NX_CACHE=true && nx serve psy-quiz-frontend",
     "lint": "nx run-many --all --target=lint",
     "fix": "npx pretty-quick --staged",
     "fix-everything": "prettier --write apps",
@@ -86,11 +86,12 @@
     "@types/jest": "29.5.13",
     "@types/node": "~18.16.9",
     "@types/passport-jwt": "^4.0.1",
-    "@typescript-eslint/eslint-plugin": "7.18.0",
-    "@typescript-eslint/parser": "7.18.0",
-    "@typescript-eslint/utils": "^7.16.0",
+    "@typescript-eslint/eslint-plugin": "8.7.0",
+    "@typescript-eslint/parser": "8.7.0",
+    "@typescript-eslint/utils": "^8.7.0",
     "angular-eslint": "^18.3.0",
-    "eslint": "^8.57.1",
+    "@eslint/js": "^9.11.1",
+    "eslint": "^9.11.1",
     "eslint-config-prettier": "^9.0.0",
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.30.0",
