@@ -1,16 +1,32 @@
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClient, HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SocketIoService } from '@services';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { AuthInterceptor, HttpErrorInterceptor } from './interceptors';
 import { AuthRestService, AuthService } from './modules';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [AppRoutingModule, BrowserModule, BrowserAnimationsModule, HttpClientModule],
-  providers: [HttpClient, AuthRestService, AuthService, SocketIoService],
+  providers: [
+    HttpClient,
+    AuthRestService,
+    AuthService,
+    SocketIoService,
+    {
+      provide: HTTP_INTERCEPTORS,
+      useClass: HttpErrorInterceptor,
+      multi: true
+    },
+    {
+      provide: HTTP_INTERCEPTORS,
+      multi: true,
+      useClass: AuthInterceptor
+    }
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
