@@ -1,3 +1,4 @@
-export * from './run-list'
-export * from './test-list'
-export * from './user-list'
+export * from './run-list';
+export * from './test-list';
+export * from './user-list';
+export * from './dashboard';
