@@ -1,4 +1,5 @@
-import { Component, inject } from '@angular/core';
+import { ChangeDetectorRef, Component, DestroyRef, inject } from '@angular/core';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { Observable } from 'rxjs';
 import { SocketIoService, UserService } from '../../../../services';
 
@@ -8,15 +9,21 @@ import { SocketIoService, UserService } from '../../../../services';
   styleUrls: ['./user-list.component.scss']
 })
 export class UserListComponent {
+  public users: any[] = [];
   private userService = inject(UserService);
   // private socketIoService = inject(SocketIoService);
-
-  public get users$(): Observable<any> {
-    return this.userService.entities$;
-  }
+  private destroyRef = inject(DestroyRef);
+  private cdr = inject(ChangeDetectorRef);
 
   constructor() {
-    this.userService.getAll().subscribe();
-    // this.socketIoService.getOnlineStatuses().subscribe((status) => {});
+    this.userService.entities$.subscribe((users) => {
+      this.users = users;
+      this.cdr.markForCheck();
+    });
+    this.userService.getAll().pipe(takeUntilDestroyed(this.destroyRef)).subscribe();
+  }
+
+  remove(testId: string): void {
+    this.userService.remove(testId).pipe(takeUntilDestroyed(this.destroyRef)).subscribe();
   }
 }
